standalone netstandard 2.0 build: removing MathF deps, use double as instead float

+++ UnityPython.BackEnd/src/Serialization/SimpleJson.UnityPython.cs
@@ -114,7 +114,7 @@ namespace SimpleJSON
                 }
                 case JSONNodeType.Float:
                 {
-                    return MK.Float(node.AsFloat);
+                    return MK.Float(node.AsDouble);
                 }
                 case JSONNodeType.NullValue:
                 {
